1-1

+++ lab2/01.py
@@ -122,7 +122,7 @@ for i in range(1,config['Epochs']+1):
         print('{}-1'.format(i))
         x, label = x.to(device ,dtype = torch.long), y.to(device ,dtype = torch.long)
         print('{}-2'.format(i))
-        pred = model(x.float())
+        pred = model(x)
         print('{}-3'.format(i))
 #         print(pred.shape)
         accuracy += torch.max(pred,1)[1].eq(label).sum().item()
